requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
cryptography==44.0.0
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.17.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.1
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
snowflake-connector-python==3.12.4
snowflake-snowpark-python==1.26.0
sortedcontainers==2.4.0
streamlit==1.41.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
watchdog==6.0.0
